requirements.txt
fastapi
uvicorn[standard]
pillow
pytesseract
sympy
transformers
torch
